Support PF_LOGGING_LEVEL, refine sdk logger usage
