+++ src/promptflow/promptflow/_cli/_pf_azure/_connection.py
@@ -2,7 +2,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 
-import logging
 from pathlib import Path
 
 from dotenv import dotenv_values
@@ -10,10 +9,11 @@ from dotenv import dotenv_values
 from promptflow._cli._params import add_param_connection_name, add_param_env, logging_params
 from promptflow._cli._utils import _set_workspace_argument_for_subparsers, activate_action, get_client_for_cli
 from promptflow._sdk._constants import LOGGER_NAME
+from promptflow._utils.logger_utils import LoggerFactory
 from promptflow.connections import CustomConnection
 from promptflow.contracts.types import Secret
 
-logger = logging.getLogger(LOGGER_NAME)
+logger = LoggerFactory.get_logger(LOGGER_NAME)
 
 
 def add_connection_parser(subparsers):
