+++ src/promptflow/promptflow/_sdk/_utils.py
@@ -5,13 +5,12 @@
 import collections
 import hashlib
 import json
-import logging
 import multiprocessing
 import os
 import platform
 import re
-import stat
 import shutil
+import stat
 import sys
 import tempfile
 import zipfile
@@ -63,6 +62,7 @@ from promptflow._sdk._errors import (
 from promptflow._sdk._vendor import IgnoreFile, get_ignore_file, get_upload_files_from_folder
 from promptflow._utils.context_utils import _change_working_dir, inject_sys_path
 from promptflow._utils.dataclass_serializer import serialize
+from promptflow._utils.logger_utils import LoggerFactory
 from promptflow.contracts.tool import ToolType
 
 
@@ -328,7 +328,7 @@ def override_connection_config_with_environment_variable(connections: Dict[str, 
     'CUSTOM_CONNECTION_CHAT_DEPLOYMENT_NAME' by default. If the environment variable is not set, it will use the
     original value as a fallback.
     """
-    logger = logging.getLogger(LOGGER_NAME)
+    logger = LoggerFactory.get_logger(LOGGER_NAME)
     for connection_name, connection in connections.items():
         values = connection.get("value", {})
         for key, val in values.items():
@@ -474,7 +474,7 @@ def _resolve_folder_to_compress(base_path: Path, include: str, dst_path: Path) -
 @contextmanager
 def _merge_local_code_and_additional_includes(code_path: Path):
     # TODO: unify variable names: flow_dir_path, flow_dag_path, flow_path
-    logger = logging.getLogger(LOGGER_NAME)
+    logger = LoggerFactory.get_logger(LOGGER_NAME)
 
     def additional_includes_copy(src, relative_path, target_dir):
         if src.is_file():
@@ -538,7 +538,7 @@ def print_pf_version():
     print("promptflow\t\t\t {}".format(get_promptflow_sdk_version()))
     print()
     print("Executable '{}'".format(os.path.abspath(sys.executable)))
-    print('Python ({}) {}'.format(platform.system(), sys.version))
+    print("Python ({}) {}".format(platform.system(), sys.version))
 
 
 class PromptflowIgnoreFile(IgnoreFile):
@@ -586,7 +586,7 @@ def _generate_tool_meta(
     *,
     include_errors_in_output: bool = False,
 ) -> Dict[str, dict]:
-    logger = logging.getLogger(LOGGER_NAME)
+    logger = LoggerFactory.get_logger(LOGGER_NAME)
     # use multi process generate to avoid system path disturb
     manager = multiprocessing.Manager()
     tools_dict = manager.dict()
