+++ src/promptflow/promptflow/_sdk/operations/_local_storage_operations.py
@@ -31,7 +31,7 @@ from promptflow._sdk.entities import Run
 from promptflow._sdk.entities._flow import Flow
 from promptflow._utils.dataclass_serializer import serialize
 from promptflow._utils.exception_utils import PromptflowExceptionPresenter
-from promptflow._utils.logger_utils import LogContext
+from promptflow._utils.logger_utils import LogContext, LoggerFactory
 from promptflow._utils.multimedia_utils import get_file_reference_encoder
 from promptflow.contracts.multimedia import Image
 from promptflow.contracts.run_info import FlowRunInfo
@@ -42,7 +42,7 @@ from promptflow.exceptions import UserErrorException
 from promptflow.executor._result import BatchResult
 from promptflow.storage import AbstractRunStorage
 
-logger = logging.getLogger(LOGGER_NAME)
+logger = LoggerFactory.get_logger(LOGGER_NAME)
 
 RunInputs = NewType("RunInputs", Dict[str, List[Any]])
 RunOutputs = NewType("RunOutputs", Dict[str, List[Any]])
