+++ src/promptflow/promptflow/azure/operations/_run_operations.py
@@ -40,12 +40,12 @@ from promptflow._sdk._constants import (
     RunStatus,
 )
 from promptflow._sdk._errors import RunNotFoundError, RunOperationParameterError
-from promptflow._sdk._logger_factory import LoggerFactory
 from promptflow._sdk._utils import in_jupyter_notebook, incremental_print
 from promptflow._sdk.entities import Run
 from promptflow._telemetry.activity import ActivityType, monitor_operation
 from promptflow._telemetry.telemetry import WorkspaceTelemetryMixin
 from promptflow._utils.flow_utils import get_flow_lineage_id
+from promptflow._utils.logger_utils import LoggerFactory
 from promptflow.azure._constants._flow import (
     AUTOMATIC_RUNTIME,
     AUTOMATIC_RUNTIME_NAME,
