+++ src/promptflow/tests/sdk_cli_azure_test/unittests/test_pf_client.py
@@ -55,12 +55,12 @@ class TestPFClient:
 
     def test_local_azure_connection_extract_workspace(self):
         res = LocalAzureConnectionOperations._extract_workspace(
-            "azureml:/subscriptions/123/resourceGroups/456/providers/Microsoft.MachineLearningServices/workspaces/789"
+            "azureml://subscriptions/123/resourceGroups/456/providers/Microsoft.MachineLearningServices/workspaces/789"
         )
         assert res == ("123", "456", "789")
 
         res = LocalAzureConnectionOperations._extract_workspace(
-            "azureml:/subscriptions/123/resourcegroups/456/workspaces/789"
+            "azureml://subscriptions/123/resourcegroups/456/workspaces/789"
         )
         assert res == ("123", "456", "789")
 
@@ -70,6 +70,6 @@ class TestPFClient:
 
         with pytest.raises(ValueError) as e:
             LocalAzureConnectionOperations._extract_workspace(
-                "azureml:/subscriptions/123/resourceGroups/456/providers/Microsoft.MachineLearningServices/workspaces/"
+                "azureml://subscriptions/123/resourceGroups/456/providers/Microsoft.MachineLearningServices/workspaces/"
             )
         assert "Malformed connection provider string" in str(e.value)
